requirements.txt
Django>=3.2.12
djangorestframework==3.13.1
stripe>=2.65.0